requirements.txt
# Study Helper v2 — Requirements

# ── Core ──────────────────────────────────────────────────────────────────────
streamlit>=1.52.2
python-dotenv>=1.2.1

# ── Supabase ──────────────────────────────────────────────────────────────────
supabase>=2.4.0

# ── LangChain ────────────────────────────────────────────────────────────────
langchain>=1.2.0
langchain-community>=0.4.1
langchain-openai>=1.1.6
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-anthropic>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# ── LLM SDKs ─────────────────────────────────────────────────────────────────
openai>=2.14.0
anthropic>=0.28.0
google-generativeai>=0.7.0

# ── Document loaders ─────────────────────────────────────────────────────────
pypdf>=6.5.0
docx2txt>=0.9
python-pptx>=1.0.2

# ── OCR ───────────────────────────────────────────────────────────────────────
mistralai>=1.0.0
Pillow>=10.4.0

# ── Web search ────────────────────────────────────────────────────────────────
tavily-python>=0.7.17

# ── TTS + Video ───────────────────────────────────────────────────────────────
gTTS>=2.5.0
moviepy==1.0.3
imageio==2.34.0
imageio-ffmpeg>=0.4.9
numpy>=1.26.0

# ── Voice (optional — better than gTTS) ──────────────────────────────────────
# elevenlabs>=1.0.0

# ── Diagrams ─────────────────────────────────────────────────────────────────
matplotlib>=3.9.0

# ── HTTP ─────────────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Dev tools ────────────────────────────────────────────────────────────────
black>=25.12.0
isort>=7.0.0